Create Expected Output Format
Analysis of Results
1.This code will print out the solutions and the performance metrics for each algorithm. The output will include:

2.The final state (list of column positions of queens in each row) for each algorithm.
3.The total number of nodes expanded during the search.
The maximum length of the queue/stack structure during the search.
The values of X and Y will be the actual metrics obtained from running the algorithms. This will give a clear comparison of the efficiency and effectiveness of BFS, DFS, and IDS for solving the 8 Queens problem.